requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
fastapi==0.109.2
frozenlist==1.4.1
future==1.0.0
idna==3.6
line-bot-sdk==3.9.0
multidict==6.0.5
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
